[api] Return pre-signed S3 url for images.
