+++ cli/tests/pcluster/api/controllers/test_image_operations_controller.py
@@ -52,7 +52,7 @@ def _create_image_info(image_id):
             "Tags": [
                 {"Key": "parallelcluster:image_id", "Value": image_id},
                 {"Key": "parallelcluster:version", "Value": "3.0.0"},
-                {"Key": "parallelcluster:build_config", "Value": "test_url"},
+                {"Key": "parallelcluster:build_config", "Value": "s3://bucket/key"},
             ],
         }
     )
@@ -67,7 +67,7 @@ def _create_stack(image_id, status, reason=None):
         "Tags": [
             {"Key": "parallelcluster:image_id", "Value": image_id},
             {"Key": "parallelcluster:version", "Value": "3.0.0"},
-            {"Key": "parallelcluster:build_config", "Value": "test_url"},
+            {"Key": "parallelcluster:build_config", "Value": "s3://bucket/key"},
             {"Key": "parallelcluster:build_log", "Value": f"arn:{image_id}:build_log"},
         ],
     }
@@ -745,6 +745,10 @@ class TestDescribeImage:
             "pcluster.aws.ec2.Ec2Client.describe_image_by_id_tag",
             return_value=_create_image_info("image1"),
         )
+        mocker.patch(
+            "pcluster.aws.s3.S3Client.create_presigned_url",
+            return_value="https://parallelcluster.aws.com/bucket/key",
+        )
 
         expected_response = {
             "creationTime": "2021-04-12T00:00:00Z",
@@ -757,11 +761,11 @@ class TestDescribeImage:
                 "tags": [
                     {"key": "parallelcluster:image_id", "value": "image1"},
                     {"key": "parallelcluster:version", "value": "3.0.0"},
-                    {"key": "parallelcluster:build_config", "value": "test_url"},
+                    {"key": "parallelcluster:build_config", "value": "s3://bucket/key"},
                 ],
             },
             "imageBuildStatus": ImageBuildStatus.BUILD_COMPLETE,
-            "imageConfiguration": {"url": "test_url"},
+            "imageConfiguration": {"url": "https://parallelcluster.aws.com/bucket/key"},
             "imageId": "image1",
             "region": "us-east-1",
             "version": "3.0.0",
@@ -782,9 +786,13 @@ class TestDescribeImage:
             "pcluster.aws.cfn.CfnClient.describe_stack",
             return_value=_create_stack("image1", CloudFormationStackStatus.CREATE_IN_PROGRESS),
         )
+        mocker.patch(
+            "pcluster.aws.s3.S3Client.create_presigned_url",
+            return_value="https://parallelcluster.aws.com/bucket/key",
+        )
 
         expected_response = {
-            "imageConfiguration": {"url": "test_url"},
+            "imageConfiguration": {"url": "https://parallelcluster.aws.com/bucket/key"},
             "imageId": "image1",
             "imageBuildStatus": ImageBuildStatus.BUILD_IN_PROGRESS,
             "cloudformationStackStatus": CloudFormationStackStatus.CREATE_IN_PROGRESS,
@@ -794,7 +802,7 @@ class TestDescribeImage:
             "cloudformationStackTags": [
                 {"Key": "parallelcluster:image_id", "Value": "image1"},
                 {"Key": "parallelcluster:version", "Value": "3.0.0"},
-                {"Key": "parallelcluster:build_config", "Value": "test_url"},
+                {"Key": "parallelcluster:build_config", "Value": "s3://bucket/key"},
                 {"Key": "parallelcluster:build_log", "Value": "arn:image1:build_log"},
             ],
             "region": "us-east-1",
@@ -824,6 +832,10 @@ class TestDescribeImage:
             "pcluster.aws.imagebuilder.ImageBuilderClient.get_image_state",
             return_value={"status": ImageBuilderImageStatus.FAILED, "reason": "img test reason"},
         )
+        mocker.patch(
+            "pcluster.aws.s3.S3Client.create_presigned_url",
+            return_value="https://parallelcluster.aws.com/bucket/key",
+        )
 
         expected_response = {
             "cloudformationStackArn": "arn:image1",
@@ -832,13 +844,13 @@ class TestDescribeImage:
             "cloudformationStackTags": [
                 {"Key": "parallelcluster:image_id", "Value": "image1"},
                 {"Key": "parallelcluster:version", "Value": "3.0.0"},
-                {"Key": "parallelcluster:build_config", "Value": "test_url"},
+                {"Key": "parallelcluster:build_config", "Value": "s3://bucket/key"},
                 {"Key": "parallelcluster:build_log", "Value": "arn:image1:build_log"},
             ],
             "cloudformationStackStatus": CloudFormationStackStatus.CREATE_FAILED,
             "cloudformationStackStatusReason": "cfn test reason",
             "imageBuildStatus": ImageBuildStatus.BUILD_FAILED,
-            "imageConfiguration": {"url": "test_url"},
+            "imageConfiguration": {"url": "https://parallelcluster.aws.com/bucket/key"},
             "imageId": "image1",
             "imagebuilderImageStatus": ImageBuilderImageStatus.FAILED,
             "imagebuilderImageStatusReason": "img test reason",
